pyproject.toml
[project]
name = "hello-cdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "aws-cdk-lib==2.185.0",
    "constructs>=10.0.0,<11.0.0",
    "boto3>=1.37.24",
    "cdk-dynamo-table-view==0.2.488",
    "cdk-ssm-documents>=0.0.12",
]

[dependency-groups]
dev = [
    "bandit>=1.8.3",
    "black>=25.1.0",
    "coverage>=7.6.10",
    "isort>=6.0.0",
    "mypy-boto3-dynamodb>=1.37.12",
    "pyright>=1.1.392.post0",
    "pytest>=8.3.4",
    "ruff>=0.9.2",
]
lint = [
    "bandit>=1.8.3",
    "isort>=6.0.0",
    "prospector>=1.13.3",
    "pyright>=1.1.392.post0",
    "ruff>=0.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = [
    "hello_cdk",
    "tests",
]

[tool.black]
target-version = ["py311", "py312", "py313"]
line-length = 100

[tool.flake8]
max-line-length = 100
max-complexity = 10



[tool.isort]
atomic = true
force_grid_wrap = 0
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
multi_line_output = 3
use_parentheses = true
